Fixed crash condition in telescope wizard autocan. Using KUndoStack instead of K3Command in the FITS Viewer. Simplified LX200 Basic layout.
The QA Wiki is great. I added a few tests and will hopefully finish most
of the list soon.

CCMAIL:kstars-devel@kde.org 


svn path=/trunk/KDE/kdeedu/kstars/; revision=709610

+++ kstars/widgets/fitshistogramdraw.cpp
@@ -107,7 +107,7 @@ void histDrawArea::paintEvent(QPaintEvent *event)
 void histDrawArea::mouseMoveEvent ( QMouseEvent * event )
 {
 
-	int event_x = event->x() - CIRCLE_DIM / 2.;
+	int event_x = (int) (event->x() - CIRCLE_DIM / 2.);
 
  	if (event->buttons() & Qt::LeftButton)
 	{
